backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
nltk==3.8.1
httpx==0.27.0
aiofiles==23.2.1
